+++ src/window/WindowCanvas.cpp
@@ -168,13 +168,13 @@ static float mapRange(float value, float rangeStart, float rangeEnd) {
 float calcPointsDistance(const ImVec2& a, const ImVec2& b) {
     float dx = b.x - a.x;
     float dy = b.y - a.y;
-    return std::sqrtf(dx * dx + dy * dy);
+    return std::sqrt(dx * dx + dy * dy);
 }
 
 // In degrees.
 static ImVec2 rotateVec2(const ImVec2& v, float angle, const ImVec2& origin) {
-    const float s = std::sinf(angle * ((float)M_PI / 180.f));
-    const float c = std::cosf(angle * ((float)M_PI / 180.f));
+    const float s = std::sin(angle * ((float)M_PI / 180.f));
+    const float c = std::cos(angle * ((float)M_PI / 180.f));
 
     float vx = v.x - origin.x;
     float vy = v.y - origin.y;
@@ -649,8 +649,8 @@ void WindowCanvas::Update() {
 
             float theta = -key.transform.angle * (float)M_PI / 180.f;
 
-            float rotatedX = scaledOffset.x * std::cosf(theta) - scaledOffset.y * std::sinf(theta);
-            float rotatedY = scaledOffset.x * std::sinf(theta) + scaledOffset.y * std::cosf(theta);
+            float rotatedX = scaledOffset.x * std::cos(theta) - scaledOffset.y * std::sin(theta);
+            float rotatedY = scaledOffset.x * std::sin(theta) + scaledOffset.y * std::cos(theta);
 
             float offsX = rotatedX / key.transform.scale.x;
             float offsY = rotatedY / key.transform.scale.y;
@@ -777,7 +777,7 @@ void WindowCanvas::Update() {
 
             // Snap to 45deg if Shift is held.
             if (shiftHeld)
-                partsTransformation.rotation = roundf(partsTransformation.rotation / 45.f) * 45.f;
+                partsTransformation.rotation = std::round(partsTransformation.rotation / 45.f) * 45.f;
         } break;
 
         default:
@@ -832,8 +832,8 @@ void WindowCanvas::Update() {
             float dx = part.transform.position.x - pX;
             float dy = part.transform.position.y - pY;
 
-            float cosAngle = std::cosf(partsTransformation.rotation * ((float)M_PI / 180.f));
-            float sinAngle = std::sinf(partsTransformation.rotation * ((float)M_PI / 180.f));
+            float cosAngle = std::cos(partsTransformation.rotation * ((float)M_PI / 180.f));
+            float sinAngle = std::sin(partsTransformation.rotation * ((float)M_PI / 180.f));
 
             float rotatedX = dx * cosAngle - dy * sinAngle;
             float rotatedY = dx * sinAngle + dy * cosAngle;
@@ -946,7 +946,7 @@ void WindowCanvas::Update() {
                     x < mCanvasSize.x;
                     x += GRID_STEP
                 ) {
-                    bool centered = std::fabsf((mCanvasTopLeft.x + x) - origin.x) < .01f;
+                    bool centered = std::fabs((mCanvasTopLeft.x + x) - origin.x) < .01f;
                     drawList->AddLine(
                         { mCanvasTopLeft.x + x, mCanvasTopLeft.y },
                         { mCanvasTopLeft.x + x, canvasBottomRight.y },
@@ -959,7 +959,7 @@ void WindowCanvas::Update() {
                     y < mCanvasSize.y;
                     y += GRID_STEP
                 ) {
-                    bool centered = std::fabsf((mCanvasTopLeft.y + y) - origin.y) < .01f;
+                    bool centered = std::fabs((mCanvasTopLeft.y + y) - origin.y) < .01f;
                     drawList->AddLine(
                         { mCanvasTopLeft.x, mCanvasTopLeft.y + y },
                         { canvasBottomRight.x, mCanvasTopLeft.y + y },
@@ -1124,7 +1124,7 @@ void WindowCanvas::Update() {
 
             constexpr float epsilon = 1e-2f;
 
-            if ((std::fabsf(partsTransformation.pivotX) > epsilon || partsTransformation.pivotY > epsilon) && !partsTransformation.pivotIsWorld) {
+            if ((std::fabs(partsTransformation.pivotX) > epsilon || partsTransformation.pivotY > epsilon) && !partsTransformation.pivotIsWorld) {
                 constexpr uint32_t pivotColorOld = IM_COL32(57, 57, 57, .3f * 255);
                 ImVec2 displayPivotPointOld = {
                     (partsAnmSpaceCenter.x * mState.zoomFactor) + origin.x,
